ci: run benchmarks before attempting to deploy them to github pages
The 'script' key in the deploy section that is supposed to run the
benchmarks isn't doing anything. Because of this, Travis bumps its head
when trying to deploy a non-existing directory to github pages, which
makes that our master branch is now consistently red.

This patch restores the setup that worked before.

This partially reverts commit aa66ef98ea6144abd2d2714fef1dfc8046e01775.