Remove some non-standard web API constructors (#2970)
This removes the EventListener, EventInit and CustomEventInit constructors from the userland globals. The type exports stay.

I removed the internal classes as well. EventListener's implementation seemed to be doing some bookkeeping on handled events but that's not being used anywhere so I assume it's old debug stuff. The other two are completely redundant.